+++ docs/app/documentation/component-docs/modal/examples/backdrop-examples/backdrop-examples.component.ts
@@ -22,7 +22,6 @@ export class BackdropExamplesComponent {
 
     openCustomBackdrop(modal): void {
         this.modalService.open(modal, {
-            container: document.activeElement as HTMLElement,
             backdropClass: 'modal-custom-overlay-example',
             backdropClickCloseable: false,
             maxWidth: '400px',