+++ docs/app/documentation/utilities/api-files.ts
@@ -37,7 +37,7 @@ export const API_FILES = {
     ],
     buttonGroup: [
         'ButtonGroupComponent',
-        'ButtonGroupedDirective'
+        'ButtonGroupedDirective',
     ],
     calendar: [
         'CalendarComponent',
@@ -119,6 +119,7 @@ export const API_FILES = {
         'ModalTitleDirective',
         'ModalCloseButtonDirective',
         'ModalConfig',
+        'ModalPosition',
         'ModalBodyComponent',
         'ModalHeaderComponent',
         'ModalFooterComponent'
@@ -179,6 +180,11 @@ export const API_FILES = {
         'SideNavigationSublinkComponent',
         'SideNavigationTitleComponent'
     ],
+    splitButton: [
+        'SplitButtonComponent',
+        'SplitButtonMenuDirective',
+        'SplitButtonActionTitle',
+    ],
     table: [
         'TableDirective',
         'ColumnSortableDirective'
@@ -186,7 +192,9 @@ export const API_FILES = {
     tabs: [
         'TabPanelComponent',
         'TabListComponent',
-        'TabTitleDirective'
+        'TabTitleDirective',
+        'TabLinkDirective',
+        'TabNavDirective'
     ],
     tile: [
         'ProductTileComponent',