+++ library/src/lib/combobox/combobox-item.ts
@@ -1,0 +1,5 @@
+export interface ComboboxItem {
+    item: any;
+    index: number;
+}
+